feat(init): initial commit with base skeleton project
