+++ backend/install.sh
@@ -1,0 +1,4 @@
+python -m venv .venv 
+source ./.venv/bin/activate # ubuntu venv path (default)
+source ./.venv/bin/Scripts/Activate # windows venv path (change as needed) 
+python -m pip install -r ./requirements.txt
