+++ backend/requirements.txt
@@ -1,2 +1,5 @@
-fastapi
+fastapi[standard]
 uvicorn
+pymilvus
+langchain
+SQLAlchemy
